style(tests,tools): apply the mechanical clang-tidy readability fixes
Behaviour-preserving fixes surfaced by widening the lint scope (#296):
uppercase integer-literal suffixes, explicit parentheses around mixed
* / + arithmetic, one declaration per statement, redundant parentheses and
duplicate prototypes removed, concise #ifdef/#elifdef, widening casts moved
onto the operands of a multiplication, and the s_ prefix on file- and
function-scope statics.

Three renames are hand-made because the auto-fix would have picked the
wrong answer: the jpeg segment fixtures are file-scope statics and take
s_, not the g_ that GlobalConstantPrefix would apply; t_png_cursor becomes
png_cursor_t; and the streamed-JPEG local renamed to s_src would have
collided with the file-scope s_src byte buffer, so it is s_pull_src.

Names fixed by an external ABI keep their spelling under a per-site NOLINT
that states why: _sbrk (newlib), _GNU_SOURCE (glibc feature test), and the
six mock-interposition seams whose macro name must equal the production
symbol it replaces -- renaming one silently un-hooks the mock and the test
would pass against the real driver.

Two real defects fell out: board_sim called cs_disasm with sizeof(code)
on an array parameter, handing capstone the pointer width instead of the
four valid instruction bytes; and tools/vela/vela_gen.py emitted an
unprefixed file-scope static, so the generator and its committed output
both now emit s_.

Refs #296

+++ tests/support/fs_fat_mount_test_util.h
@@ -127,7 +127,7 @@ static inline ra8_err_t mc_read(void* ctx, uint32_t lba, uint32_t count, uint8_t
   if (lba + count > d->block_count) {
     return k_ra8_err_out_of_range;
   }
-  memcpy(buf, &d->bytes[lba * (uint32_t)k_mc_blk], count * (uint32_t)k_mc_blk);
+  memcpy(buf, &d->bytes[(size_t)lba * (uint32_t)k_mc_blk], (size_t)count * (uint32_t)k_mc_blk);
   return k_ra8_ok;
 }
 
@@ -137,7 +137,7 @@ static inline ra8_err_t mc_write(void* ctx, uint32_t lba, uint32_t count, const 
   if (lba + count > d->block_count) {
     return k_ra8_err_out_of_range;
   }
-  memcpy(&d->bytes[lba * (uint32_t)k_mc_blk], buf, count * (uint32_t)k_mc_blk);
+  memcpy(&d->bytes[(size_t)lba * (uint32_t)k_mc_blk], buf, (size_t)count * (uint32_t)k_mc_blk);
   return k_ra8_ok;
 }
